+++ tests/test_generate.py
@@ -1,6 +1,4 @@
 #
-# -*- coding: utf-8 -*-
-#
 # Copyright (c) 2023 Jared Crapo
 #
 # Permission is hereby granted, free of charge, to any person obtaining a copy
@@ -25,10 +23,28 @@
 # pylint: disable=missing-module-docstring, unused-variable
 
 import pytest
-import rich.style
 import rich.errors
+import rich.style
 
+import shell_themer.generators
 from shell_themer import Themer
+
+
+#
+# test GeneratorBase functionality
+#
+def test_generator_classmap():
+    classmap = shell_themer.generators.GeneratorBase.classmap
+    assert "environment_variables" in classmap
+    assert "bogusgenerator" not in classmap
+    assert classmap["environment_variables"].__name__ == "EnvironmentVariables"
+
+
+def test_generator_name():
+    envgen = shell_themer.generators.EnvironmentVariables(None, None, None)
+    assert envgen.generator == "environment_variables"
+    fzfgen = shell_themer.generators.Fzf(None, None, None)
+    assert fzfgen.generator == "fzf"
 
 
 #
@@ -123,26 +139,26 @@ def test_generate_no_scopes(thm_cmdline, capsys):
 
 
 #
-# test rendering of elements common to all scopes
+# test elements common to all scopes
 #
 def test_generate_enabled(thm_cmdline, capsys):
     tomlstr = """
         [scope.nolistvar]
         enabled = false
         generator = "environment_variables"
-        environment.unset = "NOLISTVAR"
+        unset = "NOLISTVAR"
 
         [scope.somevar]
         enabled = true
         generator = "environment_variables"
-        environment.unset = "SOMEVAR"
+        unset = "SOMEVAR"
     """
     exit_code = thm_cmdline("generate", tomlstr)
     out, err = capsys.readouterr()
     assert exit_code == Themer.EXIT_SUCCESS
     assert not err
     assert "unset SOMEVAR" in out
-    assert not "unset NOLISTVAR" in out
+    assert "unset NOLISTVAR" not in out
 
 
 def test_generate_enabled_false_enabled_if_ignored(thm_cmdline, capsys):
@@ -151,7 +167,7 @@ def test_generate_enabled_false_enabled_if_ignored(thm_cmdline, capsys):
         enabled = false
         enabled_if = "[[ 1 == 1 ]]"
         generator = "environment_variables"
-        environment.unset = "NOLISTVAR"
+        unset = "NOLISTVAR"
     """
     exit_code = thm_cmdline("generate", tomlstr)
     out, err = capsys.readouterr()
@@ -166,7 +182,7 @@ def test_generate_enabled_true_enabed_if_ignored(thm_cmdline, capsys):
         enabled = true
         enabled_if = "[[ 0 == 1 ]]"
         generator = "environment_variables"
-        environment.unset = "NOLISTVAR"
+        unset = "NOLISTVAR"
     """
     exit_code = thm_cmdline("generate", tomlstr)
     out, err = capsys.readouterr()
@@ -180,7 +196,7 @@ def test_generate_enabled_invalid_value(thm_cmdline, capsys):
         [scope.unset]
         enabled = "notaboolean"
         generator = "environment_variables"
-        environment.unset = "NOLISTVAR"
+        unset = "NOLISTVAR"
     """
     exit_code = thm_cmdline("generate", tomlstr)
     out, err = capsys.readouterr()
@@ -204,12 +220,12 @@ def test_generate_enabled_if(cmd, enabled, thm_cmdline, capsys):
     tomlstr = f"""
         [variables]
         echocmd = "/bin/echo"
-        falsetest = "[[ 1 == 0]]"
+        falsetest = "[[ 1 == 0 ]]"
 
         [scope.unset]
         enabled_if = "{cmd}"
         generator = "environment_variables"
-        environment.unset = "ENVVAR"
+        unset = "ENVVAR"
     """
     exit_code = thm_cmdline("generate", tomlstr)
     out, err = capsys.readouterr()
@@ -226,12 +242,12 @@ def test_generate_comments(thm_cmdline, capsys):
         [scope.nolistvar]
         enabled = false
         generator = "environment_variables"
-        environment.unset = "NOLISTVAR"
+        unset = "NOLISTVAR"
 
         [scope.somevar]
         enabled = true
         generator = "environment_variables"
-        environment.unset = "SOMEVAR"
+        unset = "SOMEVAR"
     """
     exit_code = thm_cmdline("generate --comment", tomlstr)
     out, err = capsys.readouterr()
@@ -240,14 +256,14 @@ def test_generate_comments(thm_cmdline, capsys):
     assert "# [scope.nolistvar]" in out
     assert "# [scope.somevar]" in out
     assert "unset SOMEVAR" in out
-    assert not "unset NOLISTVAR" in out
+    assert "unset NOLISTVAR" not in out
 
 
 def test_unknown_generator(thm_cmdline, capsys):
     tomlstr = """
         [scope.myprog]
         generator = "mrfusion"
-        environment.unset = "SOMEVAR"
+        unset = "SOMEVAR"
     """
     exit_code = thm_cmdline("generate", tomlstr)
     _, err = capsys.readouterr()
@@ -264,7 +280,7 @@ def test_no_generator(thm_cmdline, capsys):
     exit_code = thm_cmdline("generate", tomlstr)
     _, err = capsys.readouterr()
     assert exit_code == Themer.EXIT_ERROR
-    assert "does not have a generator defined" in err
+    assert "does not have a generator" in err
     assert "myscope" in err
 
 
@@ -275,12 +291,12 @@ ENV_INTERPOLATIONS = [
     ("{style:dark_orange}", "#ff6c1c"),
     ("{style:dark_orange:hex}", "#ff6c1c"),
     ("{style:dark_orange:hexnohash}", "ff6c1c"),
-    # for an unknown format or style, don't do any replacement
-    ("{style:current_line}", "{style:current_line}"),
-    ("{style:dark_orange:unknown}", "{style:dark_orange:unknown}"),
+    (
+        "{style:dark_orange:ansi_on}hello there{style:dark_orange:ansi_off}",
+        "\x1b[38;2;255;108;28mhello there\x1b[0m",
+    ),
     # we have to have the style keyword, or it all just gets passed through
     ("{dark_orange}", "{dark_orange}"),
-    ("{variable:green}", "{variable:green}"),
     # escaped opening bracket, becasue this is toml, if you want a backslash
     # you have to you \\ because toml strings can contain escape sequences
     (r"\\{style:bright_blue}", "{style:bright_blue}"),
@@ -306,7 +322,7 @@ def test_generate_environment_interpolation(thm_cmdline, capsys, phrase, interpo
 
         [scope.gum]
         generator = "environment_variables"
-        environment.export.GUM_OPTS = " --cursor-foreground={phrase}"
+        export.GUM_OPTS = " --cursor-foreground={phrase}"
     """
     exit_code = thm_cmdline("generate", tomlstr)
     out, err = capsys.readouterr()
@@ -316,11 +332,14 @@ def test_generate_environment_interpolation(thm_cmdline, capsys, phrase, interpo
 
 def test_generate_environment_unset_list(thm_cmdline, capsys):
     tomlstr = """
+        [variables]
+        thevar = "ANOTHERVAR"
+
         [scope.ls]
         generator = "environment_variables"
         # set some environment variables
-        environment.unset = ["SOMEVAR", "ANOTHERVAR"]
-        environment.export.LS_COLORS = "ace ventura"
+        unset = ["SOMEVAR", "{var:thevar}"]
+        export.LS_COLORS = "ace ventura"
     """
     exit_code = thm_cmdline("generate", tomlstr)
     out, err = capsys.readouterr()
@@ -335,7 +354,7 @@ def test_generate_environment_unset_string(thm_cmdline, capsys):
     tomlstr = """
         [scope.unset]
         generator = "environment_variables"
-        environment.unset = "NOLISTVAR"
+        unset = "NOLISTVAR"
     """
     exit_code = thm_cmdline("generate", tomlstr)
     out, err = capsys.readouterr()
@@ -361,9 +380,10 @@ ATTRIBS_TO_FZF = [
 
 
 @pytest.mark.parametrize("styledef, fzf", ATTRIBS_TO_FZF)
-def test_fzf_attribs_from_style(thm, styledef, fzf):
+def test_fzf_attribs_from_style(styledef, fzf):
     style = rich.style.Style.parse(styledef)
-    assert fzf == thm._fzf_attribs_from_style(style)
+    genny = shell_themer.generators.Fzf(None, None, None, None, None)
+    assert fzf == genny._fzf_attribs_from_style(style)
 
 
 STYLE_TO_FZF = [
@@ -386,13 +406,17 @@ STYLE_TO_FZF = [
 
 
 @pytest.mark.parametrize("name, styledef, fzf", STYLE_TO_FZF)
-def test_fzf_from_style(thm, name, styledef, fzf):
+def test_fzf_from_style(name, styledef, fzf):
     style = rich.style.Style.parse(styledef)
-    assert fzf == thm._fzf_from_style(name, style)
+    genny = shell_themer.generators.Fzf(None, None, None, None, None)
+    assert fzf == genny._fzf_from_style(name, style)
 
 
-def test_fzf_opts(thm_cmdline, capsys):
+def test_fzf(thm_cmdline, capsys):
     tomlstr = """
+        [styles]
+        purple = "#7060eb"
+
         [variables]
         bstyle = "rounded"
 
@@ -401,18 +425,25 @@ def test_fzf_opts(thm_cmdline, capsys):
         environment_variable = "QQQ"
         opt."+i" = true
         opt.--border = "{var:bstyle}"
+        style.prompt = "magenta3"
+        style.info = "purple"
     """
     exit_code = thm_cmdline("generate", tomlstr)
     out, err = capsys.readouterr()
     assert exit_code == Themer.EXIT_SUCCESS
     assert not err
-    assert out == """export QQQ=" +i --border='rounded'"\n"""
+    expected = (
+        """export QQQ=" +i --border='rounded'"""
+        """ --color='prompt:164:regular,info:#7060eb:regular'"\n"""
+    )
+    assert out == expected
 
 
 def test_fzf_no_opts(thm_cmdline, capsys):
     tomlstr = """
         [variables]
         varname = "ZZ"
+
         [scope.fzf]
         generator = "fzf"
         environment_variable = "Q{var:varname}QQ"
@@ -469,9 +500,12 @@ STYLE_TO_LSCOLORS = [
 
 
 @pytest.mark.parametrize("name, styledef, expected", STYLE_TO_LSCOLORS)
-def test_ls_colors_from_style(thm, name, styledef, expected):
+def test_ls_colors_from_style(name, styledef, expected):
     style = rich.style.Style.parse(styledef)
-    code, render = thm._ls_colors_from_style(name, style, thm.LS_COLORS_MAP, "scope")
+    genny = shell_themer.generators.LsColors(None, None, None, None, None)
+    code, render = genny.ls_colors_from_style(
+        name, style, genny.LS_COLORS_MAP, prog="prog", scope="scope"
+    )
     assert render == expected
     assert code == expected[0:2]
 
@@ -513,6 +547,26 @@ def test_ls_colors_environment_variable(thm_cmdline, capsys):
     assert exit_code == Themer.EXIT_SUCCESS
     assert not err
     assert out == 'export OTHER_LS_COLOR="fi=0"\n'
+
+
+def test_ls_colors_styles_variables(thm_cmdline, capsys):
+    tomlstr = """
+        [variables]
+        pinkvar = "magenta3"
+
+        [styles]
+        warning = "yellow on red"
+
+        [scope.lsc]
+        generator = "ls_colors"
+        style.file = "warning"
+        style.directory = "{var:pinkvar}"
+    """
+    exit_code = thm_cmdline("generate", tomlstr)
+    out, err = capsys.readouterr()
+    assert exit_code == Themer.EXIT_SUCCESS
+    assert not err
+    assert out == 'export LS_COLORS="fi=33;41:di=38;5;164"\n'
 
 
 def test_ls_colors_clear_builtin(thm_cmdline, capsys):
@@ -580,9 +634,12 @@ STYLE_TO_EXACOLORS = [
 
 
 @pytest.mark.parametrize("name, styledef, expected", STYLE_TO_EXACOLORS)
-def test_exa_colors_from_style(thm, name, styledef, expected):
+def test_exa_colors_from_style(name, styledef, expected):
     style = rich.style.Style.parse(styledef)
-    code, render = thm._ls_colors_from_style(name, style, thm.EXA_COLORS_MAP, "scope")
+    genny = shell_themer.generators.ExaColors(None, None, None, None, None)
+    code, render = genny.ls_colors_from_style(
+        name, style, genny.EXA_COLORS_MAP, prog="prog", scope="scope"
+    )
     assert render == expected
     assert code == expected[0:2]
 
@@ -627,6 +684,26 @@ def test_exa_colors_environment_variable(thm_cmdline, capsys):
     assert out == 'export OTHER_EXA_COLOR="fi=0:sn=38;2;112;96;235"\n'
 
 
+def test_exa_colors_styles_variables(thm_cmdline, capsys):
+    tomlstr = """
+        [variables]
+        pinkvar = "magenta3"
+
+        [styles]
+        warning = "yellow on red"
+
+        [scope.lsc]
+        generator = "exa_colors"
+        style.file = "warning"
+        style.directory = "{var:pinkvar}"
+    """
+    exit_code = thm_cmdline("generate", tomlstr)
+    out, err = capsys.readouterr()
+    assert exit_code == Themer.EXIT_SUCCESS
+    assert not err
+    assert out == 'export EXA_COLORS="fi=33;41:di=38;5;164"\n'
+
+
 def test_exa_colors_clear_builtin(thm_cmdline, capsys):
     tomlstr = """
         [scope.exac]
@@ -659,14 +736,151 @@ def test_exa_colors_clear_builtin_not_boolean(thm_cmdline, capsys):
 
 
 #
+# test the eza_colors generator
+#
+# we only reallly have to test that the style name maps to the right code in ls_colors
+# ie directory -> di, or setuid -> su. The ansi codes are created by rich.style
+# so we don't really need to test much of that
+STYLE_TO_EZACOLORS = [
+    ("text", "", ""),
+    ("text", "default", "no=0"),
+    ("file", "default", "fi=0"),
+    ("directory", "#8be9fd", "di=38;2;139;233;253"),
+    ("symlink", "green4 bold", "ln=1;38;5;28"),
+    ("multi_hard_link", "blue on white", "mh=34;47"),
+    ("pi", "#f8f8f2 on #44475a underline", "pi=4;38;2;248;248;242;48;2;68;71;90"),
+    ("socket", "bright_white", "so=97"),
+    ("door", "bright_white", "do=97"),
+    ("block_device", "default", "bd=0"),
+    ("character_device", "black", "cd=30"),
+    ("broken_symlink", "bright_blue", "or=94"),
+    ("missing_symlink_target", "bright_blue", "mi=94"),
+    ("setuid", "bright_blue", "su=94"),
+    ("setgid", "bright_red", "sg=91"),
+    ("sticky", "blue_violet", "st=38;5;57"),
+    ("other_writable", "blue_violet italic", "ow=3;38;5;57"),
+    ("sticky_other_writable", "deep_pink2 on #ffffaf", "tw=38;5;197;48;2;255;255;175"),
+    ("executable_file", "cornflower_blue on grey82", "ex=38;5;69;48;5;252"),
+    ("file_with_capability", "red on black", "ca=31;40"),
+    ("sn", "#7060eb", "sn=38;2;112;96;235"),
+]
+
+
+@pytest.mark.parametrize("name, styledef, expected", STYLE_TO_EZACOLORS)
+def test_eza_colors_from_style(name, styledef, expected):
+    style = rich.style.Style.parse(styledef)
+    genny = shell_themer.generators.EzaColors(None, None, None, None, None)
+    code, render = genny.ls_colors_from_style(
+        name, style, genny.EZA_COLORS_MAP, prog="prog", scope="scope"
+    )
+    assert render == expected
+    assert code == expected[0:2]
+
+
+def test_eza_colors_no_styles(thm_cmdline, capsys):
+    tomlstr = """
+        [scope.exac]
+        generator = "eza_colors"
+    """
+    exit_code = thm_cmdline("generate", tomlstr)
+    out, err = capsys.readouterr()
+    assert exit_code == Themer.EXIT_SUCCESS
+    assert not err
+    assert out == 'export EZA_COLORS=""\n'
+
+
+def test_eza_colors_unknown_style(thm_cmdline, capsys):
+    tomlstr = """
+        [scope.exac]
+        generator = "eza_colors"
+        style.bundleid = "default"
+    """
+    exit_code = thm_cmdline("generate", tomlstr)
+    out, err = capsys.readouterr()
+    assert exit_code == Themer.EXIT_ERROR
+    assert "unknown style" in err
+    assert "exac" in err
+
+
+def test_eza_colors_environment_variable(thm_cmdline, capsys):
+    tomlstr = """
+        [scope.exac]
+        generator = "eza_colors"
+        environment_variable = "OTHER_EZA_COLOR"
+        style.file = "default"
+        style.size_number = "#7060eb"
+    """
+    exit_code = thm_cmdline("generate", tomlstr)
+    out, err = capsys.readouterr()
+    assert exit_code == Themer.EXIT_SUCCESS
+    assert not err
+    assert out == 'export OTHER_EZA_COLOR="fi=0:sn=38;2;112;96;235"\n'
+
+
+def test_eza_colors_styles_variables(thm_cmdline, capsys):
+    tomlstr = """
+        [variables]
+        pinkvar = "magenta3"
+
+        [styles]
+        warning = "yellow on red"
+
+        [scope.lsc]
+        generator = "eza_colors"
+        style.file = "warning"
+        style.directory = "{var:pinkvar}"
+    """
+    exit_code = thm_cmdline("generate", tomlstr)
+    out, err = capsys.readouterr()
+    assert exit_code == Themer.EXIT_SUCCESS
+    assert not err
+    assert out == 'export EZA_COLORS="fi=33;41:di=38;5;164"\n'
+
+
+def test_eza_colors_clear_builtin(thm_cmdline, capsys):
+    tomlstr = """
+        [scope.exac]
+        generator = "eza_colors"
+        clear_builtin = true
+        style.directory = "bright_blue"
+        style.uu = "bright_red"
+        style.punctuation = "#555555"
+    """
+    exit_code = thm_cmdline("generate", tomlstr)
+    out, err = capsys.readouterr()
+    assert exit_code == Themer.EXIT_SUCCESS
+    assert not err
+    expected = 'export EZA_COLORS="reset:di=94:uu=91:xx=38;2;85;85;85"\n'
+    assert out == expected
+
+
+def test_eza_colors_clear_builtin_not_boolean(thm_cmdline, capsys):
+    tomlstr = """
+        [scope.exac]
+        generator = "eza_colors"
+        clear_builtin = "error"
+        style.directory = "bright_blue"
+    """
+    exit_code = thm_cmdline("generate", tomlstr)
+    out, err = capsys.readouterr()
+    assert exit_code == Themer.EXIT_ERROR
+    assert not out
+    assert "'clear_builtin' to be true or false" in err
+
+
+#
 # test the iterm generator
 #
-def test_iterm(thm_cmdline, capsys):
+def test_iterm_fg_bg(thm_cmdline, capsys):
     tomlstr = """
+        [styles]
+        foreground = "#ffeebb"
+        background = "#221122"
+
         [scope.iterm]
         generator = "iterm"
-        style.foreground = "#ffeebb"
-        style.background = "#221122"
+        style.foreground = "foreground"
+        style.background = "background"
     """
     exit_code = thm_cmdline("generate", tomlstr)
     out, err = capsys.readouterr()
@@ -674,11 +888,11 @@ def test_iterm(thm_cmdline, capsys):
     assert not err
     lines = out.splitlines()
     assert len(lines) == 2
-    assert lines[0] == r'builtin echo -e "\e]1337;SetColors=fg=ffeebb\a"'
-    assert lines[1] == r'builtin echo -e "\e]1337;SetColors=bg=221122\a"'
+    assert lines[0] == r'builtin echo -en "\e]1337;SetColors=fg=ffeebb\a"'
+    assert lines[1] == r'builtin echo -en "\e]1337;SetColors=bg=221122\a"'
 
 
-def test_iterm_bgonly(thm_cmdline, capsys):
+def test_iterm_bg(thm_cmdline, capsys):
     tomlstr = """
         [scope.iterm]
         generator = "iterm"
@@ -690,11 +904,139 @@ def test_iterm_bgonly(thm_cmdline, capsys):
     assert not err
     lines = out.splitlines()
     assert len(lines) == 1
-    assert lines[0] == r'builtin echo -e "\e]1337;SetColors=bg=b2cacd\a"'
+    assert lines[0] == r'builtin echo -en "\e]1337;SetColors=bg=b2cacd\a"'
+
+
+def test_iterm_profile(thm_cmdline, capsys):
+    tomlstr = """
+        [scope.iterm]
+        generator = "iterm"
+        cursor = "box"
+        style.cursor = "#b2cacd"
+        profile = "myprofilename"
+    """
+    exit_code = thm_cmdline("generate", tomlstr)
+    out, err = capsys.readouterr()
+    assert exit_code == Themer.EXIT_SUCCESS
+    assert not err
+    lines = out.splitlines()
+    # we have multiple directives in this scope, but the profile directive
+    # should always come out first
+    assert len(lines) == 3
+    assert lines[0] == r'builtin echo -en "\e]1337;SetProfile=myprofilename\a"'
+
+
+def test_iterm_cursor(thm_cmdline, capsys):
+    tomlstr = """
+        [scope.iterm]
+        generator = "iterm"
+        cursor = "underline"
+        style.cursor = "#cab2cd"
+    """
+    exit_code = thm_cmdline("generate", tomlstr)
+    out, err = capsys.readouterr()
+    assert exit_code == Themer.EXIT_SUCCESS
+    assert not err
+    lines = out.splitlines()
+    assert len(lines) == 2
+    assert lines[0] == r'builtin echo -en "\e]1337;CursorShape=2\a"'
+    assert lines[1] == r'builtin echo -en "\e]1337;SetColors=curbg=cab2cd\a"'
+
+
+CURSOR_SHAPES = [
+    ("block", "0"),
+    ("box", "0"),
+    ("vertical_bar", "1"),
+    ("vertical", "1"),
+    ("bar", "1"),
+    ("pipe", "1"),
+    ("underline", "2"),
+]
+
+
+@pytest.mark.parametrize("name, code", CURSOR_SHAPES)
+def test_iterm_cursor_shape(thm_cmdline, capsys, name, code):
+    tomlstr = f"""
+        [scope.iterm]
+        generator = "iterm"
+        cursor = "{name}"
+    """
+    exit_code = thm_cmdline("generate", tomlstr)
+    out, err = capsys.readouterr()
+    assert exit_code == Themer.EXIT_SUCCESS
+    assert not err
+    lines = out.splitlines()
+    assert len(lines) == 1
+    # fr'...' lets us use f string interpolation, but the r disables
+    # escape processing, just what we need for this test
+    assert lines[0] == rf'builtin echo -en "\e]1337;CursorShape={code}\a"'
+
+
+def test_iterm_cursor_profile(thm_cmdline, capsys):
+    tomlstr = """
+        [scope.iterm]
+        generator = "iterm"
+        profile = "smoov"
+        cursor = "profile"
+    """
+    exit_code = thm_cmdline("generate", tomlstr)
+    out, err = capsys.readouterr()
+    assert exit_code == Themer.EXIT_SUCCESS
+    assert not err
+    lines = out.splitlines()
+    assert len(lines) == 2
+    # fr'...' lets us use f string interpolation, but the r disables
+    # escape processing, just what we need for this test
+    assert lines[1] == r'builtin echo -en "\e[0q"'
+
+
+def test_iterm_cursor_shape_invalid(thm_cmdline, capsys):
+    tomlstr = """
+        [scope.iterm]
+        generator = "iterm"
+        cursor = "ibeam"
+    """
+    exit_code = thm_cmdline("generate", tomlstr)
+    out, err = capsys.readouterr()
+    assert exit_code == Themer.EXIT_ERROR
+    assert not out
+    assert "unknown cursor" in err
+
+
+def test_item_tab_default(thm_cmdline, capsys):
+    tomlstr = """
+        [scope.iterm]
+        generator = "iterm"
+        style.tab = "default"
+    """
+    exit_code = thm_cmdline("generate", tomlstr)
+    out, err = capsys.readouterr()
+    assert exit_code == Themer.EXIT_SUCCESS
+    assert not err
+    lines = out.splitlines()
+    assert len(lines) == 1
+    assert lines[0] == r'builtin echo -en "\e]6;1;bg;*;default\a"'
+
+
+def test_iterm_tab_color(thm_cmdline, capsys):
+    tomlstr = """
+        [scope.iterm]
+        generator = "iterm"
+        style.tab = "#337799"
+    """
+    exit_code = thm_cmdline("generate", tomlstr)
+    out, err = capsys.readouterr()
+    assert exit_code == Themer.EXIT_SUCCESS
+    assert not err
+    lines = out.splitlines()
+    assert len(lines) == 3
+    assert lines[0] == r'builtin echo -en "\e]6;1;bg;red;brightness;51\a"'
+    assert lines[1] == r'builtin echo -en "\e]6;1;bg;green;brightness;119\a"'
+    assert lines[2] == r'builtin echo -en "\e]6;1;bg;blue;brightness;153\a"'
 
 
 #
-# test the shellcommand generator
+# test the shell generator
 #
 def test_shell(thm_cmdline, capsys):
     tomlstr = """
@@ -715,6 +1057,25 @@ def test_shell(thm_cmdline, capsys):
     assert exit_code == Themer.EXIT_SUCCESS
     assert not err
     assert out == "echo hello there\necho general kenobi\necho #a020f0\n"
+
+
+def test_shell_ansi(thm_cmdline, capsys):
+    tomlstr = """
+        [variables]
+        greeting = "hello there"
+
+        [styles]
+        purple = "#A020F0"
+
+        [scope.shortcut]
+        generator = "shell"
+        command.first = "echo {style:purple:ansi_on}{var:greeting}{style:purple:ansi_off}"
+    """  # noqa: E501
+    exit_code = thm_cmdline("generate", tomlstr)
+    out, err = capsys.readouterr()
+    assert exit_code == Themer.EXIT_SUCCESS
+    assert not err
+    assert out == "echo \x1b[38;2;160;32;240mhello there\x1b[0m\n"
 
 
 def test_shell_enabled_if(thm_cmdline, capsys):