+++ tests/test_themer.py
@@ -1,6 +1,4 @@
 #
-# -*- coding: utf-8 -*-
-#
 # Copyright (c) 2023 Jared Crapo
 #
 # Permission is hereby granted, free of charge, to any person obtaining a copy
@@ -28,291 +26,210 @@ import argparse
 import os
 
 import pytest
-import rich.style
 import rich.errors
+import rich.style
 
 from shell_themer import ThemeError
 
 
 #
-# test style parsing
+# test style and variable processing on initialization
 #
-def test_get_style_plain(thm):
-    style = thm.get_style("#aaff00")
-    assert isinstance(style, rich.style.Style)
-    assert style.color.name == "#aaff00"
-
-
-def test_get_style_complex(thm):
-    style = thm.get_style("bold white on red")
-    assert isinstance(style, rich.style.Style)
-    assert style.bold is True
-    assert style.color.name == "white"
-    assert style.bgcolor.name == "red"
-
-
-def test_get_style_invalid(thm):
-    with pytest.raises(rich.errors.StyleSyntaxError):
-        _ = thm.get_style("not a valid style")
-
-
-def test_get_style_interpolation(thm):
+def test_process_definition(theme):
     tomlstr = """
-        [variables]
-        qyellow = "#ffff00"
-
         [styles]
         background =  "#282a36"
         foreground =  "#f8f8f2"
         current_line =  "#f8f8f2 on #44475a"
-        warning = "{var:qyellow}"
-    """
-    thm.loads(tomlstr)
-    style = thm.get_style("current_line")
-    assert style.color.name == "#f8f8f2"
-    assert style.bgcolor.name == "#44475a"
-    style = thm.get_style("warning")
-    assert style.color.name == "#ffff00"
+        comment =  "#6272a4"
+        cyan =  "#8be9fd"
+        green =  "#50fa7b"
+        orange =  "#ffb86c"
+        pink =  "#ff79c6"
+        purple =  "#bd93f9"
+        yellow =  "#f1fa8c"
 
-
-def test_process_definition(thm):
-    tomlstr = """
         [variables]
-        replace = "{var:secondhalf}"
-        firsthalf = "#ff"
+        capture.somevar = "printf '%s' {var:replace}"
         secondhalf = "5555"
+        replace = "{var:secondhalf}"
+        firsthalf = "fred"
         myred = "{var:firsthalf}{variable:secondhalf}"
-        igreen = "{style:green}"
-
-        [styles]
-        background =  "#282a36"
-        foreground =  "#f8f8f2"
-        current_line =  "#f8f8f2 on #44475a"
-        comment =  "#6272a4"
-        cyan =  "#8be9fd"
-        green =  "#50fa7b"
-        orange =  "#ffb86c"
-        pink =  "#ff79c6"
-        purple =  "#bd93f9"
-        red =  "{var:myred}"
-        yellow =  "#f1fa8c"
-
-        [scope.ls]
-        # set some environment variables
-        environment.unset = ["SOMEVAR", "ANOTHERVAR"]
-        environment.export.LS_COLORS = "ace ventura"
+        igreen = "{style:green:hexnohash}"
+        capture.anothervar = "printf '%s' myvalue"
     """
-    thm.loads(tomlstr)
-    assert isinstance(thm.styles, dict)
-    assert isinstance(thm.styles["cyan"], rich.style.Style)
-    assert thm.styles["cyan"].color.name == "#8be9fd"
-    assert thm.styles["yellow"].color.name == "#f1fa8c"
-    # check an interpolated variable in a style
-    assert thm.styles["red"].color.name == "#ff5555"
-    # check the variable interpolation
-    assert thm.value_of("replace") == "5555"
-    assert thm.value_of("myred") == "#ff5555"
-    # check style interpolation in variables
-    assert thm.value_of("igreen") == "#50fa7b"
+    theme.loads(tomlstr)
+    # check the styles
+    assert isinstance(theme.styles, dict)
+    assert isinstance(theme.styles["cyan"], rich.style.Style)
+    assert theme.styles["cyan"].color.name == "#8be9fd"
+    assert theme.styles["yellow"].color.name == "#f1fa8c"
+    assert len(theme.styles) == 10
+    # check the variables
+    assert len(theme.variables) == 7
+    # capture doesn't interpolate variables
+    assert theme.variables["somevar"] == "{var:replace}"
+    # make sure capture variable actually captures
+    assert theme.variables["anothervar"] == "myvalue"
+    # styles interpolate into variables
+    assert theme.variables["igreen"] == "50fa7b"
+    # variables interpolate into variables
+    assert theme.variables["replace"] == "5555"
+    assert theme.variables["myred"] == "fred5555"
 
 
-def test_styles_from(thm):
+def test_process_definition_duplicate_variables(theme):
     tomlstr = """
-        [styles]
-        background =  "#282a36"
-        foreground =  "#f8f8f2"
-        current_line =  "#f8f8f2 on #44475a"
-        comment =  "#6272a4"
-        cyan =  "#8be9fd"
-        green =  "#50fa7b"
-        orange =  "#ffb86c"
-        pink =  "#ff79c6"
-        purple =  "#bd93f9"
-        red =  "#ff5555"
-        yellow =  "#f1fa8c"
-
-        [scope.iterm]
-        generator = "iterm"
-        style.foreground = "foreground"
-        style.background = "background"
-
-        [scope.fzf]
-        generator = "fzf"
-
-        # attributes specific to fzf
-        environment_variable = "FZF_DEFAULT_OPTS"
-
-        # command line options
-        opt.--prompt = ">"
-        opt.--border = "single"
-        opt.--pointer = "•"
-        opt.--info = "hidden"
-        opt.--no-sort = true
-        opt."+i" = true
-
-        # styles
-        style.text = "foreground"
-        style.label = "green"
-        style.border = "orange"
-        style.selected = "current_line"
-        style.prompt = "green"
-        style.indicator = "cyan"
-        style.match = "pink"
-        style.localstyle = "green on black"
+        [variables]
+        capture.thevar = "printf '%s' thevalue"
+        thevar = "fred"
     """
-    thm.loads(tomlstr)
-    scopedef = thm.scopedef_for("fzf")
-    styles = thm.styles_from(scopedef)
-    assert isinstance(styles, dict)
-    assert len(styles) == 8
-    assert "indicator" in styles.keys()
-    assert isinstance(styles["localstyle"], rich.style.Style)
-    style = styles["selected"]
-    assert style.color.name == "#f8f8f2"
-    assert style.bgcolor.name == "#44475a"
+    with pytest.raises(ThemeError):
+        theme.loads(tomlstr)
 
 
-def test_styles_from_unknown(thm):
+def test_process_definition_capture_error(theme):
+    # the extra f in printff should return a non-zero
+    # exit code, which is an error
     tomlstr = """
-        [scope.iterm]
-        generator = "iterm"
-        style.foreground = "foreground"
-        style.background = "background"
+        [variables]
+        capture.thevar = "printff '%s' thevalue"
     """
-    thm.loads(tomlstr)
-    scopedef = thm.scopedef_for("unknown")
-    styles = thm.styles_from(scopedef)
-    assert isinstance(styles, dict)
-    assert styles == {}
+    with pytest.raises(ThemeError):
+        theme.loads(tomlstr)
+
+
+def test_process_definition_undefined_variable(theme):
+    tomlstr = """
+        [variables]
+        one = "{var:two}"
+    """
+    with pytest.raises(ThemeError):
+        theme.loads(tomlstr)
+
+
+# TODO this should test the init in GeneratorBase which sets scope_styles
+# def test_styles_from(thm):
+#     tomlstr = """
+#         [styles]
+#         background =  "#282a36"
+#         foreground =  "#f8f8f2"
+#         current_line =  "#f8f8f2 on #44475a"
+#         comment =  "#6272a4"
+#         cyan =  "#8be9fd"
+#         green =  "#50fa7b"
+#         orange =  "#ffb86c"
+#         pink =  "#ff79c6"
+#         purple =  "#bd93f9"
+#         red =  "#ff5555"
+#         yellow =  "#f1fa8c"
+
+#         [scope.iterm]
+#         generator = "iterm"
+#         style.foreground = "foreground"
+#         style.background = "background"
+
+#         [scope.fzf]
+#         generator = "fzf"
+
+#         # attributes specific to fzf
+#         environment_variable = "FZF_DEFAULT_OPTS"
+
+#         # command line options
+#         opt.--prompt = ">"
+#         opt.--border = "single"
+#         opt.--pointer = "•"
+#         opt.--info = "hidden"
+#         opt.--no-sort = true
+#         opt."+i" = true
+
+#         # styles
+#         style.text = "foreground"
+#         style.label = "green"
+#         style.border = "orange"
+#         style.selected = "current_line"
+#         style.prompt = "green"
+#         style.indicator = "cyan"
+#         style.match = "pink"
+#         style.localstyle = "green on black"
+#     """
+#     thm.loads(tomlstr)
+#     scopedef = thm.scopedef_for("fzf")
+#     styles = thm.styles_from(scopedef)
+#     assert isinstance(styles, dict)
+#     assert len(styles) == 8
+#     assert "indicator" in styles.keys()
+#     assert isinstance(styles["localstyle"], rich.style.Style)
+#     style = styles["selected"]
+#     assert style.color.name == "#f8f8f2"
+#     assert style.bgcolor.name == "#44475a"
+
+
+# TODO I don't think we need to test this, as long as we test the
+# init() method of GeneratorBase
+# def test_styles_from_unknown(thm):
+#     tomlstr = """
+#         [scope.iterm]
+#         generator = "iterm"
+#         style.foreground = "foreground"
+#         style.background = "background"
+#     """
+#     thm.loads(tomlstr)
+#     scopedef = thm.scopedef_for("unknown")
+#     styles = thm.styles_from(scopedef)
+#     assert isinstance(styles, dict)
+#     assert styles == {}
 
 
 #
 # test variable related methods, including interpolation
 #
-VARIABLES = [
-    ("SomeVar", "Hello"),
-    ("another_var", "one,two,three"),
-    ("comment", "#6272a4"),
-    ("empty", ""),
-    ("number", 5),
-    ("bool", True),
-    ("notdefined", None),
-    ("text", "#6272a4 on #002200"),
-    # make sure we interpolate a style into a variable
-    ("astyle", "00ff00"),
-]
-
-
-@pytest.mark.parametrize("variable, value", VARIABLES)
-def test_value_of(thm, variable, value):
-    tomlstr = """
-        [variables]
-        SomeVar =  "Hello"
-        another_var = "one,two,three"
-        comment =  "#6272a4"
-        background = "#002200"
-        number = 5
-        bool = true
-        empty = ""
-        text = "{var:comment} on {variable:background}"
-        astyle = "{style:green:hexnohash}"
-
-        [styles]
-        green = "#00ff00"
-    """
-    thm.loads(tomlstr)
-    assert thm.value_of(variable) == value
-
-
-VARIABLE_INTERPOLATIONS = [
-    ("{variable:SomeVar} there", "Hello there"),
-    ("{variable:somevar} there", "{variable:somevar} there"),
-    ("It is {var:bool}.", "It is true."),
-    ("nothing to be done", "nothing to be done"),
-    ("fred='{var:empty}'", "fred=''"),
-    (r"\{variable:SomeVar} there", "{variable:SomeVar} there"),
-    ("I have {var:number} apples.", "I have 5 apples."),
-    ("count: {variable:another_var}", "count: one,two,three"),
-]
-
-
-@pytest.mark.parametrize("value, newvalue", VARIABLE_INTERPOLATIONS)
-def test_variable_interpolate(thm, value, newvalue):
-    tomlstr = """
-        [variables]
-        SomeVar =  "Hello"
-        another_var = "one,two,three"
-        comment =  "#6272a4"
-        number = 5
-        bool = true
-        empty = ""
-    """
-    thm.loads(tomlstr)
-    assert thm.variable_interpolate(value) == newvalue
 
 
 #
 # test scope, parsing, and validation methods
 #
-def test_scopedef(thm):
+def test_scopedef(theme):
     tomlstr = """
         [scope.iterm]
         generator = "iterm"
         style.foreground = "blue"
         style.background = "white"
     """
-    thm.loads(tomlstr)
-    scopedef = thm.scopedef_for("iterm")
+    theme.loads(tomlstr)
+    scopedef = theme.scopedef_for("iterm")
     assert isinstance(scopedef, dict)
     assert scopedef["generator"] == "iterm"
     assert len(scopedef) == 2
-    styles = thm.styles_from(scopedef)
-    assert len(styles) == 2
-    assert isinstance(styles["foreground"], rich.style.Style)
+    # TODO this should be tested on the GeneratorBase, not here
+    # styles = thm.styles_from(scopedef)
+    # assert len(styles) == 2
+    # assert isinstance(styles["foreground"], rich.style.Style)
 
 
-def test_scopedef_notfound(thm):
+def test_scopedef_notfound(theme):
     tomlstr = """
         [scope.iterm]
         generator = "iterm"
         style.foreground = "blue"
         style.background = "white"
     """
-    thm.loads(tomlstr)
-    scopedef = thm.scopedef_for("notfound")
+    theme.loads(tomlstr)
+    scopedef = theme.scopedef_for("notfound")
     assert isinstance(scopedef, dict)
     assert scopedef == {}
 
 
-def test_has_scope(thm):
+def test_has_scope(theme):
     tomlstr = """
         [scope.qqq]
         generator = "iterm"
         style.foreground = "blue"
         style.background = "white"
     """
-    thm.loads(tomlstr)
-    assert thm.has_scope("qqq")
-    assert not thm.has_scope("fred")
-
-
-BOOL_TESTS = [
-    (True, True),
-    (False, True),
-    ("something", False),
-    (0, False),
-    (1, False),
-    (0.5, False),
-]
-
-
-@pytest.mark.parametrize("val, expected", BOOL_TESTS)
-def test_assert_bool(thm, val, expected):
-    if expected:
-        thm._assert_bool(val, "generator", "scope", "key")
-    else:
-        with pytest.raises(ThemeError):
-            thm._assert_bool(val, "generator", "scope", "key")
+    theme.loads(tomlstr)
+    assert theme.has_scope("qqq")
+    assert not theme.has_scope("fred")
 
 
 #
@@ -372,8 +289,8 @@ def test_load_from_args_filename(thm, mocker, tmp_path):
     args.theme = None
 
     thm.load_from_args(args)
-    assert thm.definition
-    assert thm.styles
+    assert thm.theme.definition
+    assert thm.theme.styles
 
 
 def test_load_from_args_invalid_filename(thm, mocker, tmp_path):
@@ -414,8 +331,8 @@ def test_load_from_args_env(thm, mocker, tmp_path):
     args.theme = None
 
     thm.load_from_args(args)
-    assert thm.definition
-    assert thm.styles
+    assert thm.theme.definition
+    assert thm.theme.styles
 
 
 def test_load_from_args_env_invalid(thm, mocker, tmp_path):
@@ -449,8 +366,8 @@ def test_load_from_args_theme_file(thm, mocker, tmp_path):
     args.theme = "themefile.toml"
 
     thm.load_from_args(args)
-    assert thm.definition
-    assert thm.styles
+    assert thm.theme.definition
+    assert thm.theme.styles
 
 
 def test_load_from_args_theme_file_invalid(thm, mocker, tmp_path):
@@ -484,16 +401,16 @@ def test_load_from_args_theme_name(thm, mocker, tmp_path):
     args.theme = "themefile"
 
     thm.load_from_args(args)
-    assert thm.definition
-    assert thm.styles
+    assert thm.theme.definition
+    assert thm.theme.styles
 
 
 #
 # test loads() method
 #
-def test_loads_empty(thm):
-    thm.loads("")
-    assert isinstance(thm.definition, dict)
-    assert thm.definition == {}
-    assert isinstance(thm.styles, dict)
-    assert thm.styles == {}
+def test_loads_empty(theme):
+    theme.loads("")
+    assert isinstance(theme.definition, dict)
+    assert theme.definition == {}
+    assert isinstance(theme.styles, dict)
+    assert theme.styles == {}